requirements.txt
accelerate==1.10.1
anyio==4.11.0
bitsandbytes==0.48.1
certifi==2025.10.5
charset-normalizer==3.4.4
diffusers @ git+https://github.com/huggingface/diffusers@fa468c5d576a42f50ef467a4dc24d0326b25f656
exceptiongroup==1.3.0
filelock==3.20.0
fsspec==2025.9.0
gguf==0.17.1
h11==0.16.0
hf-xet==1.1.10
hf_transfer==0.1.9
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.35.3
idna==3.11
importlib_metadata==8.7.0
Jinja2==3.1.6
MarkupSafe==3.0.3
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.6
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.8.90
packaging==25.0
pillow==11.3.0
protobuf==6.32.1
psutil==7.1.0
PyYAML==6.0.3
regex==2025.9.18
requests==2.32.5
safetensors==0.6.2
sentencepiece==0.2.1
sniffio==1.3.1
sympy==1.14.0
tokenizers==0.22.1
torch==2.8.0
torchvision==0.23.0
tqdm==4.67.1
transformers==4.57.0
triton==3.4.0
typing_extensions==4.15.0
urllib3==2.5.0
zipp==3.23.0
